+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file
 
 ## [0.2.8] 2023-10-26
 
+### Added
+
+
+- Add more test for validate word
+
+### Fixed
+
+
+- Fixed method validate word
+
 ### Changed